Move routes configuration to separate file
